refactor: :white_check_mark: Refactored code to clearly show authors license details and attribute usage.
Code is better documented to represent where it comes from and who authored it.
